Migrate to non-misk dynamo backfila support
