+++ client-dynamodb-v2/README.md
@@ -1,0 +1,4 @@
+# client-dynamodb-v2
+
+Backfila backend client implementation to backfill a DynamoDB datastore 
+using the v2 api.
