+++ client-dynamodb-v2/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=client-dynamodb-v2
+POM_NAME=client-dynamodb-v2
+POM_DESCRIPTION=DynamoDB backfila backend client library using the Dynamo v2 API
+POM_PACKAGING=jar
