--- client-misk-dynamodb-v2/README.md
@@ -1,4 +1,0 @@
-# client-misk-dynamodb-v2
-
-Backfila Misk client implementation to backfill a DynamoDB datastore using the
-v2 api.
