--- client-misk-dynamodb-v2/gradle.properties
@@ -1,4 +1,0 @@
-POM_ARTIFACT_ID=client-misk-dynamodb-v2
-POM_NAME=client-misk-dynamodb-v2
-POM_DESCRIPTION=Misk DynamoDB client library for backfila
-POM_PACKAGING=jar
