--- client-misk-dynamodb-v2/src/test/kotlin/app/cash/backfila/client/misk/dynamodbv2/BackfillsModule.kt
@@ -1,31 +1,0 @@
-package app.cash.backfila.client.misk.dynamodbv2
-
-import app.cash.backfila.client.BackfilaClientConfig
-import app.cash.backfila.client.misk.BackfillModule
-import misk.aws2.dynamodb.DynamoDbService
-import misk.inject.KAbstractModule
-import misk.inject.toKey
-
-/**
- * Simulates a Backfills module where all the relevant backfills are registered.
- */
-class BackfillsModule : KAbstractModule() {
-  override fun configure() {
-    val dependsOn = listOf(DynamoDbService::class.toKey())
-    install(
-      BackfillModule(
-        BackfilaClientConfig(
-          url = "test.url", slack_channel = "#test"
-        ),
-        dependsOn = dependsOn
-      )
-    )
-    install(DynamoDbBackfillModule.create<DynamoDbBackfillTest.MakeTracksExplicitBackfill>())
-    install(DynamoDbBackfillModule.create<DynamoDbFilteringTest.FilteredMakeTracksExplicitBackfill>())
-    install(DynamoDbBackfillModule.create<DynamoDbFilteringTest.DynamoFilterMakeTracksExplicitBackfill>())
-    install(DynamoDbBackfillModule.create<DynamoDbSegmentingTest.SegmentingBackfill>())
-    install(DynamoDbBackfillModule.create<DynamoDbLastEvaluatedKeyTest.PausingBackfill>())
-    install(DynamoDbBackfillModule.create<DynamoDbBillingModeTest.EmptyTrackBackfill>())
-    install(DynamoDbBackfillModule.create<DynamoDbBillingModeTest.ReallyExpensiveBackfill>())
-  }
-}
